Added Saving/Loading of slider levels
Slider levels can now be saved with the 'Save Changes' button in the SliderLevelsForm
Slider levels are now loaded after the settings are loaded.
Made changes can also be reverted with the 'Revert Changes' button in the SliderLevelsForm. Only works if changes weren't saved yet.
The slider levels can be reverted to default with the 'Revert to Default' button in the SliderLevelsForm.