Default slider levels fixed
Bug:
The slider levels didn't match the default slider levels if no .config file in appdata existed.

Fix:
Changed the default value of the 'sliderLevelsRelative' setting in the project properties to match the default slider level values.

Also removed the unneeded setting 'sliderLevelsAbsolute' and removed the logic that set the values of this setting. (The setting was never loaded and thus had no effect)